+++ s2gos-server/src/s2gos_server/services/local/__init__.py
@@ -5,6 +5,7 @@
 from .job import Job, JobCancelledException, JobContext, get_job_context
 from .local_service import LocalService
 from .process_registry import ProcessRegistry
+from .registered_process import RegisteredProcess
 
 __all__ = [
     "LocalService",
@@ -12,5 +13,6 @@ __all__ = [
     "JobCancelledException",
     "JobContext",
     "ProcessRegistry",
+    "RegisteredProcess",
     "get_job_context",
 ]