+++ s2gos-server/src/s2gos_server/services/local/local_service.py
@@ -2,104 +2,67 @@
 #  Permissions are hereby granted under the terms of the Apache 2.0 License:
 #  https://opensource.org/license/apache-2-0.
 
-import traceback
 from concurrent.futures import ThreadPoolExecutor
 from concurrent.futures.process import ProcessPoolExecutor
 from typing import Any, Callable, Optional
 
 import fastapi
 import pydantic
-from starlette.routing import Route
 
 from s2gos_common.models import (
-    Capabilities,
-    ConformanceDeclaration,
     JobInfo,
     JobList,
     JobResults,
     JobStatus,
-    Link,
     ProcessDescription,
     ProcessList,
     ProcessRequest,
     ProcessSummary,
     Schema,
 )
-from s2gos_common.service import Service
 from s2gos_server.exceptions import JSONContentException
+from s2gos_server.services.base import ServiceBase
 
 from .job import Job
 from .process_registry import ProcessRegistry
-
-model_dump_config = dict(
-    exclude_none=True,
-    exclude_unset=True,
-    exclude_defaults=True,
-)
-
-html_names = ("swagger_ui_html", "swagger_ui_redirect", "redoc_html")
-
-conforms_to = [
-    "http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/core",
-    "http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/ogc-process-description",
-    "http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/json",
-    # "http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/html",
-    "http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/oas30",
-    "http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/job-list",
-    # "http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/callback",
-    "http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/dismiss",
-]
+from .registered_process import RegisteredProcess
 
 
-class LocalService(Service):
+class LocalService(ServiceBase):
     def __init__(
         self,
         title: str,
         description: Optional[str] = None,
         executor: Optional[ThreadPoolExecutor | ProcessPoolExecutor] = None,
     ):
-        self.title = title
-        self.description = description
+        super().__init__(title=title, description=description)
         self.executor = executor or ThreadPoolExecutor(max_workers=3)
         self.process_registry = ProcessRegistry()
         self.jobs: dict[str, Job] = {}
-
-    async def get_capabilities(
-        self, request: fastapi.Request, **kwargs
-    ) -> Capabilities:
-        links = _get_capabilities(request)
-        return Capabilities(
-            title=self.title,
-            description=self.description,
-            links=links,
-        )
-
-    async def get_conformance(self, **_kwargs) -> ConformanceDeclaration:
-        return ConformanceDeclaration(conformsTo=conforms_to)
 
     async def get_processes(self, request: fastapi.Request, **_kwargs) -> ProcessList:
         return ProcessList(
             processes=[
                 ProcessSummary(
-                    **p.model_dump(
+                    **p.description.model_dump(
                         mode="python",
                         exclude={"inputs", "outputs"},
                     )
                 )
-                for p in self.process_registry.get_process_list()
+                for p in self.process_registry.values()
             ],
-            links=[_get_self_link(request, "get_processes")],
+            links=[self.get_self_link(request, "get_processes")],
         )
 
     async def get_process(self, process_id: str, **kwargs) -> ProcessDescription:
-        process_entry = self._get_process_entry(process_id)
-        return process_entry.process
+        process = self._get_process(process_id)
+        return process.description
 
     async def execute_process(
         self, process_id: str, process_request: ProcessRequest, **_kwargs
     ) -> JobInfo:
-        process_entry = self._get_process_entry(process_id)
-        process_desc = process_entry.process
+        process = self._get_process(process_id)
+        process_desc = process.description
         input_params = _nest_dict(process_request.inputs or {})
         input_default_params = {
             input_name: input_info.schema_.default
@@ -116,12 +79,12 @@ class LocalService(Service):
 
         model_instance: pydantic.BaseModel
         try:
-            model_instance = process_entry.model_class(**input_values)
+            model_instance = process.model_class(**input_values)
         except pydantic.ValidationError as e:
             raise JSONContentException(
                 400,
                 detail=f"Invalid parameterization for process {process_id!r}: {e}",
-                traceback=traceback.format_exception(type(e), e, e.__traceback__),
+                exception=e,
             )
 
         function_kwargs = {
@@ -138,7 +101,7 @@ class LocalService(Service):
         job = Job(
             process_id=process_desc.id,
             job_id=job_id,
-            function=process_entry.function,
+            function=process.function,
             function_kwargs=function_kwargs,
         )
         self.jobs[job_id] = job
@@ -149,7 +112,7 @@ class LocalService(Service):
     async def get_jobs(self, request: fastapi.Request, **_kwargs) -> JobList:
         return JobList(
             jobs=[job.job_info for job in self.jobs.values()],
-            links=[_get_self_link(request, "get_jobs")],
+            links=[self.get_self_link(request, "get_jobs")],
         )
 
     async def get_job(self, job_id: str, *args, **kwargs) -> JobInfo:
@@ -181,9 +144,9 @@ class LocalService(Service):
         assert job.future is not None
         assert job.job_info.processID is not None
         result = job.future.result()
-        entry = self.process_registry.get_entry(job.job_info.processID)
-        assert entry is not None
-        outputs = entry.process.outputs or {}
+        process = self.process_registry.get(job.job_info.processID)
+        assert process is not None
+        outputs = process.description.outputs or {}
         output_count = len(outputs)
         return JobResults.model_validate(
             {
@@ -202,7 +165,7 @@ class LocalService(Service):
         input_fields: Optional[dict[str, pydantic.fields.FieldInfo]] = None,
         output_fields: Optional[dict[str, pydantic.fields.FieldInfo]] = None,
     ) -> Callable[[Callable], Callable]:
-        """A decorator for user functions to be registered as processes."""
+        """A decorator that registers a user function as a process."""
 
         def _factory(function: Callable):
             self.process_registry.register_function(
@@ -218,13 +181,13 @@ class LocalService(Service):
 
         return _factory
 
-    def _get_process_entry(self, process_id: str) -> ProcessRegistry.Entry:
-        process_entry = self.process_registry.get_entry(process_id)
-        if process_entry is None:
+    def _get_process(self, process_id: str) -> RegisteredProcess:
+        process = self.process_registry.get(process_id)
+        if process is None:
             raise JSONContentException(
                 404, detail=f"Process {process_id!r} does not exist"
             )
-        return process_entry
+        return process
 
     def _get_job(
         self, job_id: str, forbidden_status_codes: dict[JobStatus, str]
@@ -236,39 +199,6 @@ class LocalService(Service):
         if message:
             raise JSONContentException(403, detail=f"Job {job_id!r} {message}")
         return job
-
-
-def _get_self_link(request: fastapi.Request, name: str, **path_params) -> Link:
-    return Link(
-        href=str(request.url_for(name, **path_params)),
-        rel="self",
-        title=name,
-        type=_get_link_type(name),
-        hreflang="en",
-    )
-
-
-def _get_capabilities(request: fastapi.Request):
-    app: fastapi.FastAPI = request.app
-    return [_get_self_link(request, "get_capabilities")] + [
-        Link(
-            href=_get_url(request, r.path),
-            title=r.name,
-            rel="service",
-            type=_get_link_type(r.name),
-            hreflang="en",
-        )
-        for r in app.routes
-        if isinstance(r, Route)
-    ]
-
-
-def _get_link_type(name: str) -> str:
-    return "text/html" if name in html_names else "application/json"
-
-
-def _get_url(request: fastapi.Request, path: str):
-    return str(request.base_url.replace(path=path))
 
 
 def _nest_dict(flat_dict: dict[str, Any]) -> dict[str, Any]: