+++ s2gos-server/src/tests/services/local/test_process_registry.py
@@ -4,7 +4,6 @@
 
 from unittest import TestCase
 
-import pydantic
 import pytest
 from pydantic import BaseModel, Field
 from tests.helpers import BaseModelMixin
@@ -16,11 +15,7 @@ from s2gos_common.models import (
     ProcessDescription,
     Schema,
 )
-from s2gos_server.services.local import ProcessRegistry
-from s2gos_server.services.local.process_registry import (
-    create_schema_instance,
-    inline_schema_refs,
-)
+from s2gos_server.services.local import ProcessRegistry, RegisteredProcess
 
 
 def f1(x: bool, y: int) -> float:
@@ -50,11 +45,11 @@ class ProcessRegistryTest(BaseModelMixin, TestCase):
         registry = ProcessRegistry()
 
         entry = registry.register_function(f1)
-        self.assertIsInstance(entry, ProcessRegistry.Entry)
+        self.assertIsInstance(entry, RegisteredProcess)
         self.assertIs(f1, entry.function)
-        process = entry.process
+        process = entry.description
         self.assertIsInstance(process, ProcessDescription)
-        self.assertEqual("tests.services.local.test_process_registry:f1", process.id)
+        self.assertEqual("f1", process.id)
         self.assertEqual("0.0.0", process.version)
         self.assertEqual(None, process.title)
         self.assertEqual("This is f1.", process.description)
@@ -84,11 +79,11 @@ class ProcessRegistryTest(BaseModelMixin, TestCase):
         registry = ProcessRegistry()
 
         entry = registry.register_function(f2)
-        self.assertIsInstance(entry, ProcessRegistry.Entry)
+        self.assertIsInstance(entry, RegisteredProcess)
         self.assertIs(f2, entry.function)
-        process = entry.process
+        process = entry.description
         self.assertIsInstance(process, ProcessDescription)
-        self.assertEqual("tests.services.local.test_process_registry:f2", process.id)
+        self.assertEqual("f2", process.id)
         self.assertEqual("0.0.0", process.version)
         self.assertEqual(None, process.title)
         self.assertEqual("This is f2.", process.description)
@@ -139,11 +134,11 @@ class ProcessRegistryTest(BaseModelMixin, TestCase):
                 "point": Field(title="A point (x, y)"),
             },
         )
-        self.assertIsInstance(entry, ProcessRegistry.Entry)
+        self.assertIsInstance(entry, RegisteredProcess)
         self.assertIs(f2, entry.function)
-        process = entry.process
+        process = entry.description
         self.assertIsInstance(process, ProcessDescription)
-        self.assertEqual("tests.services.local.test_process_registry:f2", process.id)
+        self.assertEqual("f2", process.id)
         self.assertEqual("0.0.0", process.version)
         self.assertEqual(None, process.title)
         self.assertEqual("This is f2.", process.description)
@@ -176,11 +171,11 @@ class ProcessRegistryTest(BaseModelMixin, TestCase):
                 "y": Field(title="The Y", lt=1.0),
             },
         )
-        self.assertIsInstance(entry, ProcessRegistry.Entry)
+        self.assertIsInstance(entry, RegisteredProcess)
         self.assertIs(f2, entry.function)
-        process = entry.process
+        process = entry.description
         self.assertIsInstance(process, ProcessDescription)
-        self.assertEqual("tests.services.local.test_process_registry:f2", process.id)
+        self.assertEqual("f2", process.id)
         self.assertEqual("0.0.0", process.version)
         self.assertEqual(None, process.title)
         self.assertEqual("This is f2.", process.description)
@@ -267,9 +262,9 @@ class ProcessRegistryTest(BaseModelMixin, TestCase):
         registry = ProcessRegistry()
 
         e1 = registry.register_function(f1, id="foo", version="1.0.2", title="My Foo")
-        self.assertIsInstance(e1, ProcessRegistry.Entry)
+        self.assertIsInstance(e1, RegisteredProcess)
         self.assertIs(f1, e1.function)
-        p1 = e1.process
+        p1 = e1.description
         self.assertIsInstance(p1, ProcessDescription)
         self.assertEqual("foo", p1.id)
         self.assertEqual("1.0.2", p1.version)
@@ -284,7 +279,7 @@ class ProcessRegistryTest(BaseModelMixin, TestCase):
         entry = registry.register_function(f3, id="f3")
         self.assertEqual(
             {"point1", "point2"},
-            set(entry.process.inputs.keys()),
+            set(entry.description.inputs.keys()),
         )
         self.assertBaseModelEqual(
             InputDescription(
@@ -308,90 +303,24 @@ class ProcessRegistryTest(BaseModelMixin, TestCase):
                     }
                 ),
             ),
-            entry.process.inputs["point1"],
+            entry.description.inputs["point1"],
         )
 
     def test_register_multiple(self):
         registry = ProcessRegistry()
-
-        self.assertEqual([], registry.get_process_list())
-        self.assertEqual(None, registry.get_process("f1"))
+        self.assertEqual([], list(registry.keys()))
+        self.assertEqual(None, registry.get("f1"))
 
         registry.register_function(f1)
-        self.assertEqual(1, len(registry.get_process_list()))
-        p1 = registry.get_process_list()[0]
-        self.assertIsInstance(p1, ProcessDescription)
-        self.assertIs(p1, registry.get_process(p1.id))
+        self.assertEqual(1, len(registry))
+        p1 = list(registry.values())[0]
+        self.assertIsInstance(p1, RegisteredProcess)
+        self.assertIs(p1, registry.get(p1.description.id))
 
         registry.register_function(f2)
-        self.assertEqual(2, len(registry.get_process_list()))
-        p1, p2 = registry.get_process_list()
-        self.assertIsInstance(p1, ProcessDescription)
-        self.assertIsInstance(p2, ProcessDescription)
-        self.assertIs(p1, registry.get_process(p1.id))
-        self.assertIs(p2, registry.get_process(p2.id))
-
-    def test_create_schema_instance(self):
-        self.assertEqual(
-            Schema(type="number"),
-            create_schema_instance("x", {"type": "number"}),
-        )
-
-        with pytest.raises(pydantic.ValidationError):
-            create_schema_instance("x", {"tüp": "number"})
-
-
-class InlineRefsTest(TestCase):
-    def test_inline_schema_refs(self):
-        schema = inline_schema_refs(
-            {
-                "type": "array",
-                "items": {"$ref": "#/$defs/Line"},
-                "$defs": {
-                    "Line": {
-                        "type": "object",
-                        "properties": {
-                            "p1": {"$ref": "#/$defs/Point"},
-                            "p2": {"$ref": "#/$defs/Point"},
-                        },
-                        "required": ["p1", "p2"],
-                    },
-                    "Point": {
-                        "type": "object",
-                        "properties": {
-                            "x": {"type": "number"},
-                            "y": {"type": "number"},
-                        },
-                        "required": ["x", "y"],
-                    },
-                },
-            }
-        )
-        self.assertEqual(
-            {
-                "type": "array",
-                "items": {
-                    "type": "object",
-                    "required": ["p1", "p2"],
-                    "properties": {
-                        "p1": {
-                            "type": "object",
-                            "required": ["x", "y"],
-                            "properties": {
-                                "x": {"type": "number"},
-                                "y": {"type": "number"},
-                            },
-                        },
-                        "p2": {
-                            "type": "object",
-                            "required": ["x", "y"],
-                            "properties": {
-                                "x": {"type": "number"},
-                                "y": {"type": "number"},
-                            },
-                        },
-                    },
-                },
-            },
-            schema,
-        )
+        self.assertEqual(2, len(registry))
+        p1, p2 = registry.values()
+        self.assertIsInstance(p1, RegisteredProcess)
+        self.assertIsInstance(p2, RegisteredProcess)
+        self.assertIs(p1, registry.get(p1.description.id))
+        self.assertIs(p2, registry.get(p2.description.id))